refactor: move board stateful logic to `useBoard`
